Still tryin' to figure out what's wrong

+++ project/raytraceCuda.cuh
@@ -10,8 +10,8 @@ struct Object;
 struct Pixel;
 
 void callRaytraceKernel(double *grid, Object *objs, double numObjects,
-                        Point_Light *lightsPPM, double numLights, double Nx, 
-                        double Ny, double filmX, double filmY, 
+                        Point_Light *lightsPPM, double numLights, int Nx, 
+                        int Ny, double filmX, double filmY, 
                         double *bgColor, double *e1, double *e2, double *e3, 
                         double *lookFrom, double epsilon, double filmDepth,
                         bool antiAliased, int blockPower);
